[FIX] industry_fsm_*: project user doesn't have read access to SO
Strangely enough, this current PR that does not touch anything
project/fsm/... related makes the test `test_lot_picking` fail
deterministically on runbot, and non determinstically in local.

Even though we don't understand why this test started failing
because of our changes, it highlighted what seems to be a
meaningful issue in the code, as most operation are done
in sudo when the user belongs to the project_user group, but
not this one, where the test failed with an AccessError.

By moving one access right elevation from the override in
industry_fsm_stock to the base method in industry_fsm_sale,
this access error is avoided, and it seems like the right
solution and behavior.

closes odoo/enterprise#78876

X-original-commit: 8e118ab5892ffbc976b34e7c8bffe68ede6e5524
Related: odoo/odoo#197027
Related: odoo/documentation#11994
Signed-off-by: Antoine Vandevenne (anv) <anv@odoo.com>
Signed-off-by: anko-odoo <anko@odoo.com>