[FIX] documents_account: use the supplier currency for vendor documents
Issue:

When creating a vendor bill or Vendor Refund through the Documents module after selecting a supplier, the currency defaults to the company's currency instead of the vendor's.

However, if the supplier is selected later in the Accounting module, the correct supplier currency is applied.

Steps to reproduce:
  - in Setting > currencies > make sure the desired currencies are active
  - In Documents, upload a bill.
  - Click on the bill and assign a vendor (whose supplier currency is different from the company's currency).
  - Click on "Create Vendor Bill".

The used currency isn't that of the supplier.

opw-4406074

closes odoo/enterprise#100846

X-original-commit: 231fb45b9d6163738acb502c13d0e08ba45a021a
Signed-off-by: Florian Charlier (flch) <flch@odoo.com>
Signed-off-by: Soukaina Tyes (soty) <soty@odoo.com>