[FIX] knowledge: freeze date for knowledge_calendar_command_tour
Prior to this commit, `knowledge_calendar_command_tour` failed on the first day
of the month, if it was a Saturday. This is because it creates an `ArticleItem`
on Monday of the week containing the current date, which may be in the previous
month, and then tries to find that same `ArticleItem` in the FullCalendar
representation of the current month with week ends days hidden. In that view,
with that setting, Monday of the current week is always shown (even if it is
part of the previous month) UNLESS the first day of the month is Saturday.

This commit also fixes some selectors using `:not` pseudo-class incorrectly
(they were always true and did not verify anything).

runbot-task-114354

closes odoo/enterprise#80825

X-original-commit: 7835fe47874f6efbd8e6278da40a7c2994a82c59
Signed-off-by: Damien Abeloos (abd) <abd@odoo.com>