[FIX] helpdesk_stock: remove `suitable_product_ids` from list view
Following odoo/enterprise@07954ea8b5bccd185dc4639c5e75d5dbab8b5499,
`product_id` was removed from the list view. Therefor there is no
more reason to have `suitable_product_ids` in the list view also, as
the only reason it was present was because `product_id` had a domain
that was depending on the `suitable_product_ids`.

This should also lead to a significant performance improvement of
the list view, as `suitable_product_ids` doesn't scale well with a
larger collection of tickets than `1` which was the expected usage
in the form view.

closes odoo/enterprise#78378

Related: odoo/upgrade#7149
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>