[FIX] documents: fix list view's Select All
During sharepocalypse, we injected the ListView.Selection component
from the control panel directly instead of using t-slots.

This made our list view component limited to the features we want
to support in the kanban view, so fewer than what the webclient
supports.

This commit restores the List View features on the list view.

Task-4544692

closes odoo/enterprise#79023

X-original-commit: 1b446a2913bd6c5073f1c08515e627351ce813fe
Signed-off-by: Patrick Hoste (pko) <pko@odoo.com>
Signed-off-by: Florian Charlier (flch) <flch@odoo.com>