[IMP] pos_preparation_display: remove the ‘skip_change’ status from order lines
The “skip_change” status was introduced to ensure that the kitchen prepares order lines only when requested by the waiter.
This commit removes that status, as it will be replaced by the courses feature, which manages it more effectively, aligning with standard POS software practices.

task-4526151

closes odoo/enterprise#78418

Related: odoo/odoo#195965
Related: odoo/upgrade#7159
Signed-off-by: Joseph Caburnay (jcb) <jcb@odoo.com>