[FIX] hr_payroll_account: Fix Payment register compute
Payment register wizard fails to compute the amount, journal, etc
since the `_compute_from_lines` is not triggered. This is because
partner_id is the only field in the view to trigger that compute before
create. And when a default value is passed for the partner, there is no
need for the compute to trigger on a new record before save.

This commit removes the default partner value, and instead it is computed
directly from the line.

Part-of: odoo/enterprise#83378
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>