[FIX] account_reports: prevent error on clicking closing entry
When attempting to click 'Closing Entry' in Tax Return, an error occurs.

Steps to reproduce:
- Install the 'account_reports' module
- Create a branch in the Main Company
- Accounting > Reporting > Tax Return
- Click 'Closing Entry'

Traceback:
AttributeError: 'account.report' object has no attribute 'show_error_branch_allowed'

At [1], an error occurs because the attribute was used without calling the
method. Because at [2], the method was added in version 18.0 but later removed
in the master branch.

[1]- https://github.com/odoo/enterprise/blob/2780da762cd430f45d97ea8fa230181ccc4e30f8/account_reports/models/account_report.py#L1971
[2]- https://github.com/odoo/enterprise/pull/73018/files#diff-5fc5051f5c0211c0eec96b892e7d29e01b68d804417443502d17bccd8333d7ec

sentry-6237112729

closes odoo/enterprise#78693

X-original-commit: 2307128b6a08221dd18b3a5518eae50338f1585b
Signed-off-by: Hugo Poncelet (hupo) <hupo@odoo.com>
Signed-off-by: Aayush Modi (aamo) <aamo@odoo.com>