[IMP] *: simplify live chat consultation from ticket
*: website_helpdesk_livechat.

This PR aims to improve live chat consultations for tickets created by
a chatbot. Specifically, it enhances the formatting of the ticket
description to make the conversation easier to read, introduces a
smart button to redirect users to the discussion channel where the
ticket originated and set the ticket name according to the user request.

task-4432728

closes odoo/enterprise#76279

Related: odoo/odoo#191813
Signed-off-by: Matthieu Stockbauer (tsm) <tsm@odoo.com>