[IMP] sale_commission: improve performances by selecting records
Before this commit, the account.move and sale.order were not selected
according to the date filter of the view. As a result, we would join the
records on plans and filter them out in a second step.

This commit add some conditions to avoid loading all records directly.

Taskid: 4471762
X-original-commit: 70ec5374b76c5072c14598a8707c293efa9a462f
Part-of: odoo/enterprise#79887
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>