[FIX] helpdesk: revert hardcoded column size of ticket ref field
Before this commit, the column of ticket ref field displayed in the list
view of `helpdesk.ticket` has a size set to 30px. That hardcoded size
makes the visibility of the content inside that column harder when the
reference contains more than 6 digits.

This commit removes the hardcoded size to let the framework computes the
column size as before.

opw-4524749

closes odoo/enterprise#78601

X-original-commit: 967a61ceaf2bbc33f03c56016ba9cb79c1111827
Related: odoo/odoo#196608
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>