[IMP] industry_fsm*: include by default subtasks in main views of tasks
Before this commit, the subtasks are hidden by default when the user
goes into a project to see the tasks linked to that project. The problem
is it is difficult to notify the sub-tasks are not displayed in the view
and it is also difficult to know how to show them. The user has to
discover the new filter `Show sub-tasks` in the search view.

This commit will now display all tasks (sub-tasks included) linked to
the project to be sure the user will see all information first. If the
user would like to hide the sub-tasks then he will just have to select
the filter `Hide Sub-Tasks`.

task-4547389

closes odoo/enterprise#78523

Related: odoo/odoo#196431
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>