[IMP] iot: add Windows IoT distinction for subs
Currently, we have regular (Raspberry Pi) IoT Boxes and Widnows
("Virtual") IoT Boxes

The subscription for the Raspberry Pis cost a monthly fee to the users,
while Windows IoT is considered to be free for the users.

However when sending information about the number of IoT Boxes connected
to a database we currently don't make a difference between Windows and
Raspberry Pi IoT Boxes.

This PR adds a distinction between the two and sends only the Raspberry Pi IoT Boxes
This allows us to distinguish the two subscriptions on odoo.com

closes odoo/enterprise#83891

X-original-commit: 87afa004db3e5695ff5c43f056357085cc7a0eb7
Signed-off-by: Yaroslav Soroko (yaso) <yaso@odoo.com>
Signed-off-by: Louis Travaux (lotr) <lotr@odoo.com>