[REF] sale_timesheet_enterprise: review the way we gather leaderboard data
Before this commit, the leaderboard data are stored in each view which
means the data are not really synced between different views of
timesheets and so each time the user switches the view, the leaderboard
data has to be reloaded even if the same period is displayed.
Moreover, 2 rpcs are made to first know if the user can see the
leaderboard and then gather the leaderboard data. The first rpc is not
really useful since no data is given if the user has no access to the
leaderboard or if the feature is disabled.

This commit reviews the way the leaderboard data are fetched to only
fetch the data when needed. That is, if the grid view displays the
current period and the user switches the view to go to the view then no
rpc call should be made to fetch the leaderboard data since we display
the same period inside the timer header component.
This commit removes the extra rpc made to know if the feature is enabled
since the logic can be done in the rpc gathering the leaderboard data.
Finally, it also removes `get_billable_time_target` python method
defined in `res.company` model since that method is no longer used in
the codebase.

task-4509858

closes odoo/enterprise#78541

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>