[IMP] *external_tax*,*avatax: provide line description
We replaced the product name with the line description in stable for
Brazil [1].

Now that we're in master, we can change the account_external_tax_mixin
interface to add it properly and remove the need to browse the
line. For consistency, we also do the same for sale.order now.

[1] https://github.com/odoo/enterprise/pull/77610

task-4401787

closes odoo/enterprise#78893

Signed-off-by: Antoine Dupuis (andu) <andu@odoo.com>