[FIX] account_online_synchronization: pass required positional arguments
Currently, an exception is generated while connecting to the bank.

Error:-
"TypeError: AccountOnlineLink.create_new_bank_account_action() missing 1 required positional argument: 'journal_type'"

This is because of the latest changes were added from the below commit.

https://github.com/odoo/enterprise/pull/70591/files#diff-72ccf5e061185b2d649fb844db777d7e229c31c38e1dc68283984a0772b481bfR384

A new argument `journal_type` was added in the `create_new_bank_account_action`
method. But forget pass that required argument in the ORM call in below lines.

https://github.com/odoo/enterprise/blob/b588877a12f4952d02c747b230065c5b52f21e1a/account_online_synchronization/static/src/components/bank_configure/bank_configure.js#L68-L72

This leads to the above traceback.

sentry-6248235685

closes odoo/enterprise#78820

X-original-commit: 1a3ac149be617b5accc4a92e490317c99962dc8a
Signed-off-by: Habib Ayob (ayh) <ayh@odoo.com>
Signed-off-by: Kishan Kumar Jaiswal (kkuj) <kkuj@odoo.com>