[FIX] web_studio: tours missing fields' required flag
During testing of the "new" Chrome's headless mode (by default since
version 128), a concurrency issue arise at the end of those tours: a
last request to the server is actually sent during the unload of the
view (cf. save). This request actually reach the server even if Chrome
is already closed and results into a SQl constraint error. Indeed, the
(very basic) form views created for those tours are based on the
"res.partner" model and only display the "name" field.

This field isn't marked as "required" neither in the model, nor the view
BUT has a SQL constraint attached to it... which makes it implicitly
"required".

This commit fixes it by properly marking them as such in the arch to let
the client-side validation prevent the ultimate "save" request made
during the unload.

closes odoo/enterprise#78994

X-original-commit: ed6976f0c70de0383f8122b43c156eab15244391
Signed-off-by: Romeo Fragomeli (rfr) <rfr@odoo.com>
Signed-off-by: Pierre Paridans (app) <app@odoo.com>