[IMP] quality{,_control,_mrp}: set failure location for all control type
Commit d7cb885c93876 introduced ability to set a failure location on
`measure_on=move_line` quality points to send the failing quantity to a
temporary location in order to inspect more deeply or repair the failing
products.
This commit extends this behaviour to all `measure_on` possible values.
In case of `product`, the entire stock moves will be redirect. In case
of `operation`, the entire picking will be redirect

closes odoo/enterprise#78609

Task: 4384819
Related: odoo/upgrade#7217
Signed-off-by: William Henrotin (whe) <whe@odoo.com>