[IMP] spreadsheet: preserve original author when duplicating revisions
Before this commit, the `spreadsheet.revision` model relied on `create_uid`
to track the author of a revision. However, when copying a spreadsheet
through the version history, the revisions were duplicated, and `create_uid`
was overridden with the user initiating the copy, leading to a loss of the
original author information.

To address this, an `author` field is added to `spreadsheet.revision` to
store the original author's reference, ensuring it remains intact even when
the spreadsheet is duplicated.

task-4510201

closes odoo/enterprise#78608

Related: odoo/upgrade#7313
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>