[FIX] l10n_be_hr_payroll: wrong year holiday attest
Bug:
When doing a departure holiday attest for an employee leaving in between years,
the payslip of the current year are actually the ones from the previous year.

Reason:
We were computing the year based on the start of the notice period.

Fix:
The year is based on the end of the notice period when there is one.

closes odoo/enterprise#78661

Task: 4547648
X-original-commit: b4f32eef24c256572277d4c6e36481929610971e
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>