[FIX] account_intrastat: Load more traceback
The issue occurs because the query does not apply an offset or limit, causing the same lines to be loaded repeatedly. This results in duplicate keys on the client side.

To fix this, we add a query tail and mimic the ordering of account_move_line, ensuring consistency when loading additional lines.

This fix is related to an earlier patch in older versions, but due to the refactoring of the Intrastat report, it had to be implemented differently.

Old Fix: https://github.com/odoo/enterprise/pull/76543

closes odoo/enterprise#78851

X-original-commit: 7bb84427770e8c8b8d5c9f8c3c0cae3d0a690674
Signed-off-by: Hugo Poncelet (hupo) <hupo@odoo.com>
Signed-off-by: Logan Staelens (lost) <lost@odoo.com>