[FIX] *: tests - sub-menus & overrides
This commit allows the `defineMenus` test helper to accept a flat list
of objects instead of a hierarchy of parent/child menus when defining
mock menus.

Additionally, it allows to override previously defined menu items by
defining a menu with the same ID. For consistency, the same has been
done to actions in `defineActions` (instead of throwing an error).

Finally, it adds error handling in case a given child menu ID is missing
from the `load_menus` call.

X-original-commit: b0d2cfaaccf1d79467e95fac85ce28da0c215931
Part-of: odoo/enterprise#80824
Related: odoo/odoo#200448
Signed-off-by: Michaël Mattiello (mcm) <mcm@odoo.com>
Signed-off-by: Julien Mougenot (jum) <jum@odoo.com>