[IMP] documents: check access when pinning an action
Purpose
=======
Check that we have write access on a folder to be able to pin an action.

Create a new type of action (new `usage`). With that type, we check that the action
is pinned on the folder of the documents we are executing the action on.

Remove unused file.

Task-4535626

closes odoo/enterprise#78691

Related: odoo/odoo#203710
Signed-off-by: Patrick Hoste (pko) <pko@odoo.com>