[FIX] l10n_it_pos: Fix the computation of down payment / global discount
Fix the tax rounding issues in down payment and global discount applied on sale orders and pos orders.
To do so, this commit introduces a new field 'extra_tax_data' allowing to split the global taxes
computation on subset of lines. This is usefull on the down payment because, when generating the
final invoice, the first down payment is subtracted with negative lines. However, those lines
represent a separated document so the taxes computation in round globally has to be split between the
original lines and the negative down payment lines.
This extra field could also contain manual tax amounts to be taken into account in the taxes computation
engine.

task_id: 4457175

closes odoo/enterprise#79852

Related: odoo/odoo#198600
Signed-off-by: Antoine Dupuis (andu) <andu@odoo.com>