[FIX] delivery_usps_rest: use correct test tags
Before this commit the class `TestMockedDeliveryUSPS` used
`'standard', '-external'` which is the opposite of what it
was supposed to do. As a result the tests were run during normal
builds, and led to errors when the tests were run without demo data.

This commit also removes the reference to demo data to make
the tests more robust.

closes odoo/enterprise#79217

X-original-commit: 1ea251c8ddc8f14d4270559994534dbfb9ac7a55
Signed-off-by: William Braeckman (wbr) <wbr@odoo.com>