[FIX] account_winbooks_import: fix account types mapping
The mapping that is currently used can lead to a constraint
on accout types being triggered, blocking the import procedure.

This commit aims at fixing this by removing the account code range
for the recievable and payable types.

From now on, these will only be applied when the corresponding
'CENTRALID' matches respectively 'S' for payables
and 'C' for receivable.

task-4465663

X-original-commit: 1d13e2758d49978331eccd8b6fa1b883d2768b12
Part-of: odoo/enterprise#78784
Signed-off-by: William André (wan) <wan@odoo.com>
Signed-off-by: John Laterre (jol) <jol@odoo.com>