[IMP] l10n_in_reports: consider irn for gstr2b matching
This commit adds an `IRN number` in matching parameters/keys.
If the bills from `GSTR2B` have an IRN number then we consider the IRN number
for finding the bill in the system and matching the bill with the GSTR2B bill.

Also, we are removing ref from matching keys for false ref. A false ref can
consider any gstr2b bill with only false ref to go through a matching process
and match them partially matched.

If there is no bill found in the system, but the bill is present on GSTR2B then
we are creating the bill, here if the IRN number is available then add it to
the new bill created.

Irn number must be unique for bills so adding a constrain for unique IRN
number for bills per company.

task-4408351

closes odoo/enterprise#79041

Signed-off-by: Josse Colpaert (jco) <jco@odoo.com>