[FIX] account_accountant: save when removing analytic rec line
How to reproduce the issue:

- Create an analytic distribution model.

- In the bank reconciliation, create a new transaction. In manual operations, remove the analytic.

If you reload the reconciliation, the analytic distribution is not deleted.

In the `bank_rec_widget` of account_accountant, the method `_line_value_changed_analytic_distribution` is triggered when the analytic field is modified. However it does not save the new value of the analytic on the account.move.line.

opw-4460079

closes odoo/enterprise#82350

X-original-commit: 86a1c966d4d05c72a61be9cb5aa79a7cdfb71a96
Signed-off-by: John Laterre (jol) <jol@odoo.com>