[IMP] account_online_{synchronization,payment}: Better payment status management
This commit has 2 different purposes:

1. Connect the bank synchronization and the payment module to a
new API route to get the connection status instead of only getting
the consent expiring date.
This new route still provides the consent expiring date but will
also provides more information about the payment initiation.
Three new information are provider, whether the payment feature
is enabled, whether the payment activation is requested to the
provider and whether the feature is activated with the provider.

2. This second purpose is directly linked to the first one, it
modifies the Initiate payment visibility. Before it was visible
once the bank journal was connected to a bank synchronization, now
it's visible once the bank journal has a bank synchronization where
the payment initiation feature is enabled by the customer.

task-4532157

closes odoo/enterprise#78740

Related: odoo/upgrade#7194
Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>