[IMP] *: gantt: splitting tool does not block drag and drop
When a splitting tool is hovered, it is impossible to drag and drop the pill
because the splitting tool presence prevent some events to be fired.
Here we refactor the splitting feature by using a single component with class
pe-none that positions near the cursor position when a splitting is possible.
If a pill is clicked and the cursor position is above the splitting tool,
a splitting is done.

Task ID: 4350819

closes odoo/enterprise#78683

Signed-off-by: Julien Carion (juca) <juca@odoo.com>