[FIX] l10n_be_reports_post_wizard: noupdate of fps bank account
When updating the module, the bank data is reloaded. When the account is already trusted (`allow_out_payment == True`), the write fails due to checks to prevent fraud.
To fix this, we wrap the data in a `noupdate` to prevent further writes from happening.
If ever the account number changes, a new record will be created instead of updating the existing one.

Since this account is the default one stipulated on the official fps site, it can be trusted by default.
https://finances.belgium.be/fr/entreprises/tva/paiement-remboursement/paiement#q1

Issue reported by upgrade team.

closes odoo/enterprise#78850

X-original-commit: 50ad7b764dcf4bfb27ec1476f1fd3b35c4a3e695
Signed-off-by: de Wouters de Bouchout Jean-Benoît (jbw) <jbw@odoo.com>
Signed-off-by: Habib Ayob (ayh) <ayh@odoo.com>