[REF] sale_subscription: improve upsell description based on start date
Before this commit, When creating an upsell quote for a subscription, a note is
automatically added to the sales order showing the prorated period, but it
doesn’t account for the start date on the quote.

With this commit, The upsell description now recalculates automatically if the
upsell start date is changed.

Moreover, before this commit, _compute_discount on sale.order.line in
subscription would update line name and create the line_note. This is
considered a bad practice to create/update records in a compute when
those records are not directly related to the purpose of the compute.

_compute_discount should only compute the discount, nothing else.

This change makes sure that the discount line note is created and
updated in an override of write, which is easier to debug, more readable
and logical.

task-4299774

closes odoo/enterprise#73438

Related: odoo/upgrade#7241
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>