[FIX] website_helpdesk: fix two minor color issues
This commit fixes two minor color issues on the `website_helpdesk` form
page.

Prior to this commit, the title of the form was using a `text-muted`,
which is inconsistent with the other titles design across the front-end.

There was also an issue about the card using a `bg-secondary` utility
class to tweak its design, which doesn't work anymore with the new
secondary color we introduced in Commit[1] and highlight the fact we
can't rely on utility classes for this scenario since we don't know
what color will be set by the user.

Commit[1]: odoo/odoo@53f605399755f9355554c2582ea0f693c8f947b4

task-4498381

closes odoo/enterprise#77466

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>